Moved util methods in SplooshModel.cs to separate util file.
